jsp적용

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -13,6 +13,7 @@ import com.example.demo.vo.Article;
 import com.example.demo.vo.ResultData;
 
 import jakarta.servlet.http.HttpSession;
+
 @Controller
 public class UsrArticleController {
 
@@ -30,18 +31,23 @@ public class UsrArticleController {
 
 		}
 
-		return ResultData.from("S-1", Ut.f("%d번 게시글 입니다", id), article);
+	
+		return ResultData.from("S-1", Ut.f("%d번 게시글 입니다", id), "게시글 1개", article);
 	}
 
+	// 로그인 체크 -> 유무 체크 -> 권한 체크 -> 수정
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
 	public ResultData<Article> doModify(HttpSession httpSession, int id, String title, String body) {
+
 		boolean isLogined = false;
 		int loginedMemberId = 0;
+
 		if (httpSession.getAttribute("loginedMemberId") != null) {
 			isLogined = true;
 			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
 		}
+
 		if (isLogined == false) {
 			return ResultData.from("F-A", "로그인 하고 써");
 		}
@@ -52,22 +58,32 @@ public class UsrArticleController {
 			return ResultData.from("F-1", Ut.f("%d번 게시글은 없습니다", id));
 		}
 
+		ResultData loginedMemberCanModifyRd = articleService.loginedMemberCanModify(loginedMemberId, article);
+
+		if (loginedMemberCanModifyRd.isFail()) {
+			return loginedMemberCanModifyRd;
+		}
+
 		articleService.modifyArticle(id, title, body);
 
 		article = articleService.getArticleById(id);
 
-		return ResultData.from("S-1", Ut.f("%d번 게시글을 수정했습니다", id), article);
+		return ResultData.from(loginedMemberCanModifyRd.getResultCode(), loginedMemberCanModifyRd.getMsg(), "수정 된 게시글",
+				article);
 	}
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
 	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
+
 		boolean isLogined = false;
 		int loginedMemberId = 0;
+
 		if (httpSession.getAttribute("loginedMemberId") != null) {
 			isLogined = true;
 			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
 		}
+
 		if (isLogined == false) {
 			return ResultData.from("F-A", "로그인 하고 써");
 		}
@@ -75,23 +91,30 @@ public class UsrArticleController {
 		Article article = articleService.getArticleById(id);
 
 		if (article == null) {
-			return ResultData.from("F-1", Ut.f("%d번 게시글은 없습니다", id), id);
+			return ResultData.from("F-1", Ut.f("%d번 게시글은 없습니다", id), "입력한 id", id);
+		}
+
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", Ut.f("%d번 게시글에 대한 권한이 없습니다", id));
 		}
 
 		articleService.deleteArticle(id);
 
-		return ResultData.from("S-1", Ut.f("%d번 게시글을 삭제했습니다", id), id);
+		return ResultData.from("S-1", Ut.f("%d번 게시글을 삭제했습니다", id), "입력한 id", id);
 	}
 
 	@RequestMapping("/usr/article/doWrite")
 	@ResponseBody
 	public ResultData<Article> doWrite(HttpSession httpSession, String title, String body) {
+
 		boolean isLogined = false;
 		int loginedMemberId = 0;
+
 		if (httpSession.getAttribute("loginedMemberId") != null) {
 			isLogined = true;
 			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
 		}
+
 		if (isLogined == false) {
 			return ResultData.from("F-A", "로그인 하고 써");
 		}
@@ -109,15 +132,14 @@ public class UsrArticleController {
 
 		Article article = articleService.getArticleById(id);
 
-	
-		return ResultData.newData(writeArticleRd, article);
+		return ResultData.newData(writeArticleRd, "생성된 게시글", article);
 	}
 
 	@RequestMapping("/usr/article/getArticles")
 	@ResponseBody
 	public ResultData<List<Article>> getArticles() {
 		List<Article> articles = articleService.getArticles();
-		return ResultData.from("S-1", "Article List", articles);
+		return ResultData.from("S-1", "Article List", "게시글 목록", articles);
 	}
 
 }
